+++ nsgt/cython/nsigtf_loop.pyx
